move @add_code_examples
