+++ starknet_py/net/gateway_client.py
@@ -46,31 +46,8 @@ from starknet_py.net.schemas.gateway import (
     TypesOfTransactionsSchema,
 )
 from starknet_py.transaction_exceptions import TransactionNotReceivedError
+from starknet_py.utils.code_examples import add_code_examples
 from starknet_py.utils.sync import add_sync_methods
-from starknet_py.utils.sync.sync import T
-
-
-def add_code_examples(original_class) -> T:
-    base_class = original_class.__base__  # pyright: ignore
-    for method_name, method in original_class.__dict__.items():
-        docstring = f"""
-        .. literalinclude:: ../starknet_py/tests/e2e/docs/code_examples/test_{original_class.__name__}.py
-            :language: python
-            :start-after: docs: {method_name.strip("_")}_start
-            :end-before: docs: {method_name.strip("_")}_end
-            :dedent: 4
-        """
-
-        if (
-            callable(method)
-            and method.__doc__ is None
-            and method_name in base_class.__dict__
-        ):
-            method.__doc__ = getattr(base_class, method_name).__doc__ + docstring
-        elif callable(method):
-            method.__doc__ = (method.__doc__ or "") + docstring
-
-    return original_class
 
 
 @add_sync_methods
